Extend SetMember to handle JsonArray and JsonObject
When a member can't be deduced with reflection and the JsonPointer segment, try to handle the object as a JsonValue.
If the segment are parsed into an Int32, it will be handled as a JsonArray. Otherwise it will be handled as a JsonObject.
In cases where the incoming value are not a JsonValue, a JsonObject or a JsonArray, only a JsonValue.Null will be upserted.
In case of a JsonArray, when the index does not exists, JsonValue.Null are added until the index exists.