Add JsonValueSerializer
An IPrioritizedSerializer that handles serialization and deserialization of JsonValue.
De-/Serialization of JsonValues will ultimately be de-/serialization of the inner values.
Since inner values also can be JsonObject and JsonArray, this serializer supports those types as well.

The serializer will follow the "$ref" references rules as the other references does (relative json-pointer path in object).
Deserialization will construct the full json, so references will be replaced with the data.